+++ string_utils.h
@@ -3,10 +3,11 @@
 
 namespace rde
 {
+
 //-----------------------------------------------------------------------------
-template<typename T> int strlen(const T* str)
+template<typename T> size_t strlen(const T* str)
 {
-	int len(0);
+	size_t len(0);
 	while (*str++)
 		++len;
 	return len;
@@ -15,7 +16,7 @@ template<typename T> int strlen(const T* str)
 //-----------------------------------------------------------------------------
 template<typename T> int strcompare(const T* s1, const T* s2, size_t len)
 {
-	for (/**/; len != 0; --len)
+	for (; len != 0; --len)
 	{
 		const T c1 = *s1;
 		const T c2 = *s2;
